The training algorithm doesn't work.
The training algorithm isn't working. I'm pretty sure I'm following the
one that was given in the book pretty closely. I wrote a function to
write the data from the network training to a file, but the string that
gets built is too big and it causes everything to crash.